backend/pyproject.toml
[tool.poetry]
name = "template-gen-backend"
version = "0.1.0"
description = ""
authors = ["maheshmad <mahesh.madhusudanan@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20240706"
tqdm = "^4.66.4"
tiktoken = "^0.7.0"
nltk = "^3.8.1"
pydantic-settings = "^2.3.4"
redis = "^5.0.8"
numpy = "^2.0.1"
pydantic = "^2.8.2"
rich = "^13.7.1"
openai = "^1.40.1"
fastapi = {extras = ["standard"], version = "^0.112.0"}
motor = "^3.1.1"
sse-starlette = "^2.1.3"
pymongo = "^4.11.0"
python-dotenv = "^1.0.1"
pydantic_settings = "^2.3.4"
python_multipart = "^0.0.12"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
motor = "^3.1.1"
pymongo = "^4.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
load = "app.loader:main"
local = "app.assistants.local_assistant:main"
export = "app.export:main"